Refactor worker structure to make it easier to follow
